Removed date from description/comment body.
The date on the issue and comment is correct when using the issue
import API, so no need to also include it in the text.

I enabled trim_blocks and lstrip_blocks in jinja, in order to format
the issues a little better. I added headers to the comment depending
on what kind of comment it is.